refactored for reused settings and tools code
