+++ includes/settings/js/src/components/DashboardDisplay.jsx
@@ -1,6 +1,5 @@
 import React, { useRef, useState } from "react";
 import { sprintf, __ } from "@wordpress/i18n";
-import { dispatch } from "@wordpress/data";
 import { showError, showSuccess } from "../toasts";
 import { insertSidebarMenuItem } from "../utils";
 import { useHistory } from "react-router-dom";
@@ -10,6 +9,8 @@ import ExportFileButton from "./ExportFileButton";
 import ImportFileButton from "./ImportFileButton";
 import { Card } from "../../../../shared-assets/js/components/card";
 import LinkList from "./LinkList";
+import { saveUsageTrackingSetting } from "../settings.service";
+import * as toolService from "../tools.service";
 
 const { wp } = window;
 const { apiFetch } = wp;
@@ -88,24 +89,10 @@ export default function DashboardDisplay({ modelChartOptions, taxonomies }) {
 	 * Sets and saves tracking data for analytics
 	 * @param {*} event
 	 */
-	function saveUsageTrackingSetting(event) {
-		// @todo catch save errors
-		dispatch("core").saveSite({
-			atlas_content_modeler_usage_tracking: event.target.value,
-		});
+	function saveTrackingSetting(event) {
+		saveUsageTrackingSetting(event);
 		setUsageTracking(event.target.value);
 		showSuccess("Your setting was updated!");
-	}
-
-	/**
-	 * Gets model export data via the REST API.
-	 */
-	async function getModels() {
-		return apiFetch({
-			path: `/wpe/atlas/content-models/`,
-			method: "GET",
-			_wpnonce: wpApiSettings.nonce,
-		});
 	}
 
 	/**
@@ -115,30 +102,7 @@ export default function DashboardDisplay({ modelChartOptions, taxonomies }) {
 	 * @returns {array} Invalid properties or an empty array if no properties were invalid.
 	 */
 	function validateField(field) {
-		let invalidProperties = [];
-
-		const requiredFieldProperties = {
-			all: ["type", "id", "name", "slug"],
-			number: ["numberType"],
-			text: ["inputType"],
-			relationship: ["cardinality", "reference"],
-		};
-
-		requiredFieldProperties.all.forEach((requiredProperty) => {
-			if (!field.hasOwnProperty(requiredProperty)) {
-				invalidProperties.push(requiredProperty);
-			}
-		});
-
-		if (field?.type && requiredFieldProperties.hasOwnProperty(field.type)) {
-			requiredFieldProperties[field.type].forEach((requiredProperty) => {
-				if (!field.hasOwnProperty(requiredProperty)) {
-					invalidProperties.push(requiredProperty);
-				}
-			});
-		}
-
-		return invalidProperties;
+		toolService.validateField(field);
 	}
 
 	/**
@@ -150,35 +114,7 @@ export default function DashboardDisplay({ modelChartOptions, taxonomies }) {
 	 * @return {object} Invalid model data, or empty object if no invalid data.
 	 */
 	async function validateModel(model, storedModelData) {
-		let errors = {};
-
-		const existingModels = Object.keys(storedModelData);
-
-		if (model?.slug && existingModels.includes(model.slug)) {
-			errors["alreadyExists"] = true;
-		}
-
-		const requiredModelProperties = ["slug", "singular", "plural"];
-		requiredModelProperties.forEach((requiredProperty) => {
-			if (!model.hasOwnProperty(requiredProperty)) {
-				if (!errors.hasOwnProperty("missingProperties")) {
-					errors["missingProperties"] = [];
-				}
-				errors["missingProperties"].push(requiredProperty);
-			}
-		});
-
-		Object.entries(model?.fields || {}).forEach(([fieldId, fieldData]) => {
-			const fieldErrors = validateField(fieldData);
-			if (fieldErrors?.length > 1) {
-				if (!errors.hasOwnProperty("fieldErrors")) {
-					errors["fieldErrors"] = [];
-				}
-				errors["fieldErrors"][fieldId] = fieldErrors;
-			}
-		});
-
-		return errors;
+		toolService.validateModel(model, storedModelData);
 	}
 
 	/**
@@ -188,18 +124,7 @@ export default function DashboardDisplay({ modelChartOptions, taxonomies }) {
 	 * @return {object} Invalid models data, or empty object if no invalid data.
 	 */
 	async function validateModels(models) {
-		let errors = {};
-
-		let storedModelData = await getModels();
-
-		for (const [index, modelData] of Object.entries(models)) {
-			const modelErrors = await validateModel(modelData, storedModelData);
-			if (Object.keys(modelErrors).length > 0) {
-				errors[modelData?.slug || index] = modelErrors;
-			}
-		}
-
-		return errors;
+		toolService.validateModels(models);
 	}
 
 	/**
@@ -207,128 +132,14 @@ export default function DashboardDisplay({ modelChartOptions, taxonomies }) {
 	 * @param {*} modelErrors
 	 */
 	function showImportErrors(modelErrors) {
-		let message = "";
-
-		Object.entries(modelErrors).forEach(([modelSlug, errors]) => {
-			if (errors?.alreadyExists) {
-				message +=
-					"<li>" +
-					sprintf(
-						__(
-							// translators: The name of the model.
-							"The ‘%s’ model already exists.",
-							"atlas-content-modeler"
-						),
-						modelSlug
-					);
-				+"</li>";
-			}
-
-			if (errors?.missingProperties) {
-				message +=
-					"<li>" +
-					sprintf(
-						__(
-							// Translators: 1: The name of the model. 2: A list of missing properties, such as "id".
-							"The ‘%1$s’ model is missing properties: %2$s.",
-							"atlas-content-modeler"
-						),
-						modelSlug,
-						errors.missingProperties.join(", ")
-					) +
-					"</li>";
-			}
-
-			if (errors?.fieldErrors) {
-				let fieldMessages = "";
-
-				Object.entries(errors.fieldErrors).forEach(
-					([fieldKey, missingProperties]) => {
-						fieldMessages +=
-							"<li>" +
-							sprintf(
-								__(
-									// Translators: 1: The id of the field. 2: A list of missing properties, such as "type".
-									"Field ‘%1$s’ is missing: %2$s.",
-									"atlas-content-modeler"
-								),
-								fieldKey,
-								missingProperties.join(", ")
-							) +
-							"</li>";
-					}
-				);
-
-				message +=
-					"<li>" +
-					// Translators: 1: The name of the model. 2: A list of missing properties.
-					sprintf(
-						__(
-							"The ‘%1$s’ model has fields with missing properties:\n %2$s",
-							"atlas-content-modeler"
-						),
-						modelSlug,
-						"<ul>" + fieldMessages + "</ul>"
-					) +
-					"</li>";
-			}
-		});
-
-		showError(
-			sprintf(
-				"<strong>%1$s</strong><ul>%2$s</ul>",
-				__(
-					"Errors prevented import. No changes have been made.",
-					"atlas-content-modeler"
-				),
-				message
-			)
-		);
+		toolService.showImportErrors(modelErrors);
 	}
 
 	/**
 	 * Uploads the file data to the API.
 	 */
 	async function createModels(formData) {
-		const parsedData = JSON.parse(formData);
-		const modelData = Object.values(parsedData);
-		const modelErrors = await validateModels(modelData);
-
-		if (Object.keys(modelErrors).length !== 0) {
-			showImportErrors(modelErrors);
-			fileUploaderRef.current.value = null;
-			return;
-		}
-
-		apiFetch({
-			path: "/wpe/atlas/content-models",
-			method: "PUT",
-			_wpnonce: wpApiSettings.nonce,
-			data: modelData,
-		})
-			.then((res) => {
-				if (res.success) {
-					showSuccess(
-						__(
-							"The models were successfully imported.",
-							"atlas-content-modeler"
-						)
-					);
-
-					modelData.forEach(insertSidebarMenuItem);
-					fileUploaderRef.current.value = null;
-				}
-			})
-			.catch((err) => {
-				showError(
-					err?.message ||
-						__(
-							"An unknown import error occurred",
-							"atlas-content-modeler"
-						)
-				);
-				fileUploaderRef.current.value = null;
-			});
+		toolService.createModels(formData);
 	}
 
 	/**
@@ -721,7 +532,7 @@ export default function DashboardDisplay({ modelChartOptions, taxonomies }) {
 																!usageTracking
 															}
 															onChange={
-																saveUsageTrackingSetting
+																saveTrackingSetting
 															}
 														></input>
 														{__(
@@ -741,7 +552,7 @@ export default function DashboardDisplay({ modelChartOptions, taxonomies }) {
 															"1"
 														}
 														onChange={
-															saveUsageTrackingSetting
+															saveTrackingSetting
 														}
 													></input>
 													<label
